simplified interface, renaming
