+++ app/src/main/java/de/tum/in/tumcampusapp/component/tumui/feedback/FeedbackContract.kt
@@ -36,6 +36,5 @@ interface FeedbackContract {
         fun listenForLocation()
         fun onSaveInstanceState(outState: Bundle)
         fun detachView()
-       // fun showImageOptionsDialog()
     }
 }
